Made some bad assumptions about Labber API: this commit punts on them
Previously, I assumed that LogFile.getNumberOfEntries would return a number
equal to 1 + the largest argument for which LogFile.getEntry returns a nonempty
entry. This turns out to be false, as in some cases it returns 1 when in fact
there are many nonempty entries.

In order to avoid taking a stance on the behavior of the Labber API, entries in
the log_history are now simply Labber.LogFile objects.