Added `no_cache` parameter to @result methods.
This flag recomputes a measurement result wthout performing the measurement
again. The @result decorator now forbids the method it wraps from having
a parameter named `no_cache` in its signature. If one is found, an exception
is raised.

I like this approach enough that I would really like to adopt it for the pesky
`index` argument that the result methods have had to date.